Revert "Canceled float Order"
This reverts commit 3474c18cd7037102503d38929e360014d182bac4.

+++ src/EcoFarmDataModule/EcoFarmAdmin/EcoFarmAdmin/Migrations/ApplicationContextModelSnapshot.cs
@@ -47,8 +47,8 @@ namespace EcoFarmAdmin.Migrations
                     b.Property<string>("Commentary")
                         .HasColumnType("TEXT");
 
-                    b.Property<float>("Order")
-                        .HasColumnType("REAL");
+                    b.Property<int>("Order")
+                        .HasColumnType("INTEGER");
 
                     b.HasKey("Id");
